Task/union type image resource video preview (#136)
# *union type image resource video preview*

## :recycle: Current situation & Problem

https://github.com/StanfordSpezi/SpeziKt/pull/133#discussion_r1818065626

## :gear: Release Notes 
- added RemoteImageSource and extended ImageSourceComposeable
accordingly

## :pencil: Code of Conduct & Contributing Guidelines 

By submitting creating this pull request, you agree to follow our [Code
of
Conduct](https://github.com/StanfordSpezi/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordSpezi/.github/blob/main/CONTRIBUTING.md):
- [x] I agree to follow the [Code of
Conduct](https://github.com/StanfordSpezi/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordSpezi/.github/blob/main/CONTRIBUTING.md).

---------
